chore(deploy_docs.yml): add custom.scss to the output for debugging purposes
This change adds a `cat` command to print the contents of `quartz-build/quartz-build/quartz/styles/custom.scss` to the workflow logs. This is done to aid in debugging potential issues with the custom SCSS file during the documentation deployment process. If the file is not being applied as expected, its contents will now be visible in the workflow run output, making it easier to identify the problem.